runtime: Split FluentBundle to bundle.py to avoid circular imports
